Add files via upload
Objective of the project is to enhance customer experience and optimize sales for AeroFit, the leading fitness equipment brand, by developing a sophisticated recommendation system for its diverse range of treadmills. This project aims to not only provide immediate value in terms of improved customer recommendations but also to establish a data-driven foundation for ongoing business optimization within AeroFit.