Add files via upload
The primary objective of this project is to leverage SQL and conduct a thorough analysis of Target's orders dataset from 2016 to 2018 in Brazil. Through exploratory data analysis (EDA) and in-depth exploration, the project aims to unravel patterns, trends, and key insights to inform strategic decisions for market expansion in the Brazil region. 

Dataset Overview:
•	The project utilizes a dataset comprising 100,000 data-points of Target's orders in Brazil, spanning the years 2016 to 2018.
•	The dataset encompasses a diverse range of attributes, including order prices, freight, and other relevant parameters.