Add files via upload
The primary goal of this project is to leverage Python to conduct a comprehensive analysis of the Netflix dataset, encompassing all movies and TV-shows available on the platform. By employing basic matrices and non-graphical techniques, the project aims to provide actionable insights to assist Netflix in strategic decision-making regarding content production and business expansion in diverse markets.